# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teleassistance API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/patients": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Muestra todos los pacientes paginados con filtros de zona y ciudad",
                "operationId": "0df0d65f049161931654c3e188ca8f93",
                "parameters": [
                    {
                        "name": "zone",
                        "in": "query",
                        "description": "Filtrar por zona",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filtrar por ciudad",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pacientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PatientResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Crea un nuevo paciente",
                "operationId": "2149729f1bf1af8877352f04be3a683f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePatientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paciente creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/patients/{id}": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Muestra un paciente",
                "operationId": "900a000f9712c06fdf96fa64333bbaff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente recuperado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Patients"
                ],
                "summary": "Actualiza un paciente",
                "operationId": "49d64723ad7c62df6eddc6fa8a778755",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePatientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paciente actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Patients"
                ],
                "summary": "Elimina un paciente",
                "operationId": "0187379ab8f57fd78eeb510643be993b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Paciente eliminado con éxito"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/patients/{id}/calls": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Obtiene las llamadas de un paciente",
                "operationId": "2910606ae517db6617360588dd7be813",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de llamadas del paciente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CallResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No hay pacientes asignados en esta zona"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/patients/{id}/alerts": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Obtiene las alertas de un paciente",
                "operationId": "54cbfc687ea393aa776c10b7db09a844",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de alertas del paciente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AlertResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No hay alertas en la zona de este paciente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePatientRequest": {
                "description": "Validación para la creación de pacientes",
                "required": [
                    "dni",
                    "name",
                    "birth_date",
                    "address",
                    "phone_number",
                    "health_card",
                    "email",
                    "zone_id",
                    "personal_situation",
                    "health_situation",
                    "housing_situation",
                    "economic_situation",
                    "autonomy"
                ],
                "properties": {
                    "dni": {
                        "description": "DNI del paciente (único)",
                        "type": "string",
                        "maxLength": 255,
                        "example": "12345678A"
                    },
                    "name": {
                        "description": "Nombre del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Juan Pérez"
                    },
                    "birth_date": {
                        "description": "Fecha de nacimiento del paciente",
                        "type": "string",
                        "format": "date",
                        "example": "1985-06-15"
                    },
                    "address": {
                        "description": "Dirección del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Calle Mayor 123, Madrid"
                    },
                    "phone_number": {
                        "description": "Número de teléfono del paciente (único)",
                        "type": "string",
                        "maxLength": 255,
                        "example": "+34678901234"
                    },
                    "health_card": {
                        "description": "Número de tarjeta sanitaria (único)",
                        "type": "string",
                        "maxLength": 255,
                        "example": "HC123456789"
                    },
                    "email": {
                        "description": "Correo electrónico del paciente (único)",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "juan.perez@example.com"
                    },
                    "zone_id": {
                        "description": "ID de la zona a la que pertenece el paciente",
                        "type": "integer",
                        "example": 1
                    },
                    "associated_teleoperator_id": {
                        "description": "ID del teleoperador asociado al paciente",
                        "type": "integer",
                        "example": 1
                    },
                    "personal_situation": {
                        "description": "Situación personal del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Vive solo"
                    },
                    "health_situation": {
                        "description": "Situación de salud del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Diabetes tipo 2"
                    },
                    "housing_situation": {
                        "description": "Situación de vivienda del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Piso de alquiler"
                    },
                    "economic_situation": {
                        "description": "Situación económica del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Pensión de jubilación"
                    },
                    "autonomy": {
                        "description": "Grado de autonomía del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Autónomo con asistencia ocasional"
                    },
                    "emergency_contacts": {
                        "description": "Lista de IDs de contactos de emergencia asociados al paciente",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdatePatientRequest": {
                "description": "Validación para la actualización de un paciente",
                "required": [
                    "dni",
                    "name",
                    "birth_date",
                    "address",
                    "phone_number",
                    "health_card",
                    "email",
                    "zone_id",
                    "personal_situation",
                    "health_situation",
                    "housing_situation",
                    "economic_situation",
                    "autonomy"
                ],
                "properties": {
                    "dni": {
                        "description": "DNI del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "12345678A"
                    },
                    "name": {
                        "description": "Nombre del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Juan Pérez"
                    },
                    "birth_date": {
                        "description": "Fecha de nacimiento del paciente",
                        "type": "string",
                        "format": "date",
                        "example": "1985-06-15"
                    },
                    "address": {
                        "description": "Dirección del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Calle Mayor 123, Madrid"
                    },
                    "phone_number": {
                        "description": "Número de teléfono del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "+34678901234"
                    },
                    "health_card": {
                        "description": "Número de tarjeta sanitaria (único)",
                        "type": "string",
                        "maxLength": 255,
                        "example": "HC123456789"
                    },
                    "email": {
                        "description": "Correo electrónico del paciente",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "juan.perez@example.com"
                    },
                    "zone_id": {
                        "description": "ID de la zona a la que pertenece el paciente",
                        "type": "integer",
                        "example": 1
                    },
                    "associated_teleoperator_id": {
                        "description": "ID del teleoperador asociado al paciente",
                        "type": "integer",
                        "example": 1
                    },
                    "personal_situation": {
                        "description": "Situación personal del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Vive solo"
                    },
                    "health_situation": {
                        "description": "Situación de salud del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Diabetes tipo 2"
                    },
                    "housing_situation": {
                        "description": "Situación de vivienda del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Piso de alquiler"
                    },
                    "economic_situation": {
                        "description": "Situación económica del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Pensión de jubilación"
                    },
                    "autonomy": {
                        "description": "Grado de autonomía del paciente",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Autónomo con asistencia ocasional"
                    },
                    "emergency_contacts": {
                        "description": "Lista de IDs de contactos de emergencia asociados al paciente",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                },
                "type": "object"
            },
...
            "AlertResource": {
                "description": "Esquema del recurso Alerta",
                "properties": {
                    "id": {
                        "description": "Identificador de la alerta",
                        "type": "integer"
                    },
                    "alert": {
                        "$ref": "#/components/schemas/AlertSubtypeCompactResource"
                    },
                    "frequency": {
                        "description": "Frecuencia de la alerta",
                        "type": "string"
                    },
                    "days_of_week": {
                        "description": "Días de la semana en los que se activa la alerta",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "zone": {
                        "$ref": "#/components/schemas/ZoneResource"
                    }
                },
                "type": "object"
            },
...
            "CallResource": {
                "description": "Esquema del recurso Llamada",
                "properties": {
                    "id": {
                        "description": "Identificador de la llamada",
                        "type": "integer"
                    },
                    "teleoperator": {
                        "$ref": "#/components/schemas/TeleoperatorResource"
                    },
                    "patient": {
                        "$ref": "#/components/schemas/PatientResource"
                    },
                    "call_type": {
                        "description": "Tipo de llamada",
                        "type": "string",
                        "enum": [
                            "Entrante",
                            "Saliente"
                        ]
                    },
                    "type": {
                        "description": "Tipo específico de la llamada",
                        "type": "string"
                    },
                    "call_time": {
                        "description": "Fecha y hora de la llamada",
                        "type": "string",
                        "format": "date-time"
                    },
                    "title": {
                        "description": "Título de la llamada",
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "description": "Descripción de la llamada",
                        "type": "string",
                        "nullable": true
                    },
                    "alert": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/CallTitleResource"
                            }
                        ],
                        "nullable": true,
                        "description": "Alerta asociada a la llamada"
                    }
                },
                "type": "object"
            },
...
            "PatientResource": {
                "description": "Esquema del recurso Paciente",
                "properties": {
                    "id": {
                        "description": "Identificador del paciente",
                        "type": "integer"
                    },
                    "dni": {
                        "description": "DNI del paciente",
                        "type": "string"
                    },
                    "name": {
                        "description": "Nombre del paciente",
                        "type": "string"
                    },
                    "birth_date": {
                        "description": "Fecha de nacimiento del paciente",
                        "type": "string",
                        "format": "date"
                    },
                    "address": {
                        "description": "Dirección del paciente",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "Número de teléfono del paciente",
                        "type": "string"
                    },
                    "health_card": {
                        "description": "Número de tarjeta sanitaria del paciente",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del paciente",
                        "type": "string",
                        "format": "email"
                    },
                    "zone": {
                        "$ref": "#/components/schemas/ZoneResource"
                    },
                    "associated_teleoperator_id": {
                        "$ref": "#/components/schemas/TeleoperatorResource"
                    },
                    "personal_situation": {
                        "description": "Situación personal del paciente",
                        "type": "string"
                    },
                    "health_situation": {
                        "description": "Situación de salud del paciente",
                        "type": "string"
                    },
                    "housing_situation": {
                        "description": "Situación de vivienda del paciente",
                        "type": "string"
                    },
                    "economic_situation": {
                        "description": "Situación económica del paciente",
                        "type": "string"
                    },
                    "autonomy": {
                        "description": "Nivel de autonomía del paciente",
                        "type": "string"
                    },
                    "emergency_contacts": {
                        "description": "Contactos de emergencia asociados al paciente",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EmergencyContactResourceRecursion"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}